Modified ViewRecipe Class
Modified View Recipe to display the recipe list and display recipe details when View Recipe is pressed. And update ingriedents quanties based on scale value.

+++ FoodItem.cs
@@ -1,16 +1,16 @@
 ﻿using System;
-
+using PROG6221_FINAL;
 // Base class representing a food item
 namespace PROG6221_FINAL
 {
     public class FoodItem
     {
         public string Name { get; set; }
-        public float Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Unit { get; set; }
 
         // Constructor for the FoodItem class.
-        public FoodItem(string name, float quantity, string unit)
+        public FoodItem(string name, double quantity, string unit)
         {
             Name = name;
             Quantity = quantity;
