+++ Ingredient.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-
+using PROG6221_FINAL;
 // Derived class representing an ingredient
 namespace PROG6221_FINAL
 {
@@ -9,7 +9,7 @@ namespace PROG6221_FINAL
         public int FoodGroupIndex { get; set; } // New property
 
         // Constructor for the Ingredient class.
-        public Ingredient(string name, float quantity, string unit, int calorieCount, int foodGroupIndex) : base(name, quantity, unit)
+        public Ingredient(string name, double quantity, string unit, int calorieCount, int foodGroupIndex) : base(name, quantity, unit)
         {
             CalorieCount = calorieCount;
             FoodGroupIndex = foodGroupIndex;
